requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
yfinance==0.2.36
newsapi-python==0.2.7
python-dotenv==1.0.0
joblib==1.3.2
beautifulsoup4==4.12.2
requests==2.31.0
flask==2.3.3